CALai/backend/pyproject.toml
[project]
name = "resumematch-backend"
version = "0.1.0"
description = "ResumeMatch AI Backend — FastAPI modular monolith"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.2.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "redis>=5.0.0",
    "boto3>=1.34.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "google-generativeai>=0.7.0",
    "httpx>=0.27.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
ml = [
    "spacy>=3.7.0",
    "sentence-transformers>=3.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "coverage>=7.4.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
